コードリファクタリング
- アップロードファイルをセッション利用しない対応を別クラスへ抽出
